// apps/k-pay-app/src/types/graphql.ts
export interface UserType {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
}

export interface AdminType {
  id: string;
  firstName: string;
  lastName: string;
}

export interface MerchantType {
  id: string;
  businessName: string;
  businessType: string;
}

export interface KycApplicationType {
  id: string;
  bankInfoStatus?: string;
  contactInfoStatus?: string;
  errors?: string[] | null;
  financialInfoStatus?: string;
  identityStatus?: string;
  insertedAt: string;
  kycClientId?: string;
  kycClientType?: string;
  message?: string | null;
  personalInfoStatus?: string;
  rejectionReason?: string | null;
  status: string;
  updatedAt: string;
}

export interface UserAccountType {
  id: string;
  role: string;
  status: 'active' | 'inactive' | 'locked' | 'pending';
  walletCode?: string;
  currentSignInAt?: string;
  currentSignInIp?: string;
  lastSignInAt?: string;
  lastSignInIp?: string;
  signInCount?: number;
  signInIp?: string;
  insertedAt: string;
  updatedAt: string;
  kycApplications?: KycApplicationType[];
  user?: UserType;
  admin?: AdminType;
  merchant?: MerchantType;
}